@@ -15,8 +15,8 @@ export class DashboardComponent implements OnInit {
 
   ngOnInit() {
     this.authService.getProfile().subscribe(profile => {
+      console.log(profile);
       this.user = profile.user;
-      // console.log(this.user);
     },
       err => {
       return false;
