@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import {AuthService} from "../../services/auth.service";
 import {Router} from "@angular/router";
-import {FlashMessagesService} from "angular2-flash-messages";
+import {FlashMessagesService} from "angular2-flash-messages/module";
 
 @Component({
   selector: 'app-login',
