@@ -9,7 +9,7 @@ export class AuthGuard implements CanActivate{
     private router:Router
   ) {}
 
-  canActivate(){
+  canActivate():boolean {
     if (this.authService.loggedIn()){
       return true;
     }
